完善了midi播放的进度控制 完善移除播放中midi文件功能

+++ Ventrue/src/core/Synth/Channel.cpp
@@ -4,7 +4,7 @@
 
 namespace ventrue
 {
-	Channel::Channel(Track* track, int channelNum)
+	Channel::Channel(Track* track, uint64_t channelNum)
 	{
 		Clear();
 		this->track = track;
